add support for Chef::Mixin::shellout
