+++ lib/chefspec/errors.rb
@@ -34,6 +34,7 @@ module ChefSpec
     class DataBagNotStubbed < NotStubbed; end
     class DataBagItemNotStubbed < NotStubbed; end
     class ShellOutNotStubbed < ChefSpecError; end
+    class MixinShellOutNotStubbed < ChefSpecError; end
 
     class CookbookPathNotFound < ChefSpecError; end
     class GemLoadError < ChefSpecError; end
