+++ lib/chefspec/extensions/chef/resource.rb
@@ -64,29 +64,6 @@ module ChefSpec::Extensions::Chef::Resource
   end
 
   #
-  # Defang shell_out and friends so it can never run.
-  #
-  if ChefSpec::API::StubsFor::HAS_SHELLOUT_COMPACTED.satisfied_by?(Gem::Version.create(Chef::VERSION))
-    def shell_out_compacted(*args)
-      return super unless $CHEFSPEC_MODE
-
-      raise ChefSpec::Error::ShellOutNotStubbed.new(args: args, type: "resource", resource: self)
-    end
-
-    def shell_out_compacted!(*args)
-      return super unless $CHEFSPEC_MODE
-
-      shell_out_compacted(*args).tap(&:error!)
-    end
-  else
-    def shell_out(*args)
-      return super unless $CHEFSPEC_MODE
-
-      raise ChefSpec::Error::ShellOutNotStubbed.new(args: args, type: "resource", resource: self)
-    end
-  end
-
-  #
   # tracking
   #
